requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1
django-cors-middleware==1.3.1
django-environ==0.4.5
django-extensions==2.1.2
django-rest-swagger==2.2.0
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
pep8==1.7.1
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycodestyle==2.4.0
PyJWT==1.4.2
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
six==1.10.0
uritemplate==3.0.0
urllib3==1.23
